updated ROC and Molchan curves
Problem on the x scale are now fixed. I gave the possibility to plot figures 6 (a, b, and c) choosing from the linear or log x-axis. To make this choice is possible to set "linear==True" for linear x-axis or "linear==False" for log x-axis. the standard remain the log x-axis.